daemon-kit: ruote amqp handler: open reply_queue as durable.
ruote-amqp's receiver opens the reply queue as durable; ensure both sides open
the reply_queue with the same settings.

When running against a late RabbitMQ, when sending a response for a workitem
back to the reply_queue, the following message is received:

NOT_ALLOWED - parameters for queue 'ruote_workitems' in vhost 'foo' not
equivalent in AMQP::Protocol::Queue::Declare

Seems like root cause is that RabbitMQ 1.8 and later are getting stricter
about queue equivalence:

http://lists.rabbitmq.com/pipermail/rabbitmq-announce/2010-June/000025.html

"...honour exchange and queue equivalence requirements for AMQP 0-9-1,
especially for queue and exchange redeclaration."

Tested with RabbitMQ 2.1.1, nothing earlier.